requirements.txt
aiogram>=3.4
SQLAlchemy>=2.0
aiosqlite>=0.19
aiosmtplib>=3.0
python-dotenv>=1.0
pydantic>=2.7

# Security & Development
ruff>=0.1.5
mypy>=1.6.1
types-all>=1.0.0
bandit>=1.7.5
safety>=2.3.5
pre-commit>=3.5.0
